Move idasen_desk coordinator to separate module
